variable type of port-id in ceilometer driver is changed from integer to string
due to some type differences of port_id variable in ceilometer driver code
it is giving type error during execution and also switch_port counters were
not getting polled.

so in this commit we were type casting the port-id variable and keeping
it as integer so that it wont show any type error and also will not
effect the polling of counters.

Change-Id: I48fd2fd64054d9b9db85b5ae490234efe17be67d
Signed-off-by: Yenuganti Vasudha <yenuganti.vasudha@ericsson.com>
Closes-Bug:#1806571
Related-Prod: PROD-27299